requirements.txt
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0
dlib>=19.24.0
Pillow>=10.0.0
flask>=2.3.0
flask-cors>=4.0.0

# Optional: For advanced emotion detection models
# tensorflow>=2.13.0
# keras>=2.13.0

